+++ src/main/java/com/github/brandtjo/releasescripthelper/action/ScriptCreationCanceledException.java
@@ -1,0 +1,8 @@
+package com.github.brandtjo.releasescripthelper.action;
+
+public class ScriptCreationCanceledException extends RuntimeException {
+
+	public ScriptCreationCanceledException(String message) {
+		super(message);
+	}
+}