+++ CHANGELOG.md
@@ -1,7 +1,26 @@
 <!-- Keep a Changelog guide -> https://keepachangelog.com -->
 
-# release-script-helper Changelog
+# Release Script Helper Changelog
 
 ## [Unreleased]
-### Added
-- Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+
+## [1.2.2] - 2021-08-21
+
+- Added latest version of Github Template to set up the plugin
+- Changed to Java 11 as targeted language level
+
+## [1.2.0]
+
+- Added generation menu action: selected text will be copied to the generated release script (Hotkey: ctrl alt r)</li>
+- Changed to project-relative file path for default directory
+
+## [1.1.0]
+- Added option to not use timestamp as file prefix, but a custom script number
+- Added project settings for selection of defaults for the dialog, especially configurable ticket types and file endings
+- Added project setting for default directory
+
+## [1.0.1]
+- Fixed sql comment symbols in generated release script
+
+## [1.0.0]
+- Added initial project setup (contains only support for sql files and 3 ticket prefixes)